Fixing async issue with mocha

+++ index.js
@@ -12,20 +12,27 @@ app.use(bodyParser.urlencoded({ extended: false }));
 // Init ressources
 var api = require('./ressources').initialize(app);
 
-// test login authentification
-db.sequelize
-  .authenticate()
-  .then(function(success) {
-    console.log('Connection has been established successfully.');
-  })
-  .catch(function (err) {
-    console.log('Unable to connect to the database:', err);
-  });
+// Start DB and server
+if(process.env.NODE_ENV == 'development') {
 
-// Sync database with new models
-db.sequelize.sync({force:true});
+  // test authentification
+  db.sequelize
+    .authenticate()
+    .then(function(success) {
+      console.log('Connection has been established successfully.');
+    })
+    .catch(function (err) {
+      console.log('Unable to connect to the database:', err);
+    });
+
+  // sync DB
+  db.sequelize.sync({force:true});
+
+};
 
 // start server
 var port = process.env.PORT || 3000;
 server = http.createServer(app);
 server.listen(port);
+
+module.exports = server;
